Code-splitting to reduce app-hosted load time (BL-11515)
